+++ ledger-service/cli-opts/BUILD.bazel
@@ -19,6 +19,9 @@ da_scala_library(
     visibility = ["//visibility:public"],
     deps = [
         "//ledger/ledger-api-common",
+        "@maven//:ch_qos_logback_logback_classic",
+        "@maven//:ch_qos_logback_logback_core",
         "@maven//:io_netty_netty_handler",
+        "@maven//:org_slf4j_slf4j_api",
     ],
 )
