+++ ledger-service/cli-opts/src/main/scala/cliopts/GlobalLogLevel.scala
@@ -1,20 +1,20 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.sandbox
+package com.daml.cliopts
 
-import ch.qos.logback.classic.Level
 import org.slf4j.{Logger, LoggerFactory}
+import ch.qos.logback.classic.{Level => LogLevel}
 
 object GlobalLogLevel {
-  def set(level: Level): Unit = {
+  def set(serviceName: String)(level: LogLevel): Unit = {
     val rootLogger = LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME)
     LoggerFactory.getILoggerFactory match {
       case loggerContext: ch.qos.logback.classic.LoggerContext =>
-        rootLogger.info(s"Sandbox verbosity changed to $level")
+        rootLogger.info(s"${serviceName} verbosity changed to $level")
         loggerContext.getLoggerList.forEach(_.setLevel(level))
       case _ =>
-        rootLogger.warn(s"Sandbox verbosity cannot be set to requested $level")
+        rootLogger.warn(s"${serviceName} verbosity cannot be set to requested $level")
     }
   }
 }
