+++ ledger/sandbox-classic/src/main/scala/platform/sandbox/SandboxMain.scala
@@ -4,6 +4,7 @@
 package com.daml.platform.sandbox
 
 import com.daml.ledger.resources.ResourceContext
+import com.daml.cliopts.GlobalLogLevel
 import com.daml.platform.sandbox.config.SandboxConfig
 import com.daml.resources.ProgramResource
 
@@ -12,7 +13,7 @@ object SandboxMain {
   def main(args: Array[String]): Unit =
     new ProgramResource({
       val config: SandboxConfig = Cli.parse(args).getOrElse(sys.exit(1))
-      config.logLevel.foreach(GlobalLogLevel.set)
+      config.logLevel.foreach(GlobalLogLevel.set("Sandbox"))
       SandboxServer.owner(Name, config)
     }).run(ResourceContext.apply)
