+++ ledger/sandbox-common/BUILD.bazel
@@ -30,6 +30,7 @@ da_scala_library(
     deps = [
         "//daml-lf/data",
         "//language-support/scala/bindings",
+        "//ledger-service/cli-opts",
         "//ledger-service/jwt",
         "//ledger/caching",
         "//ledger/cli-opts",
@@ -48,12 +49,10 @@ da_scala_library(
         "//libs-scala/resources-akka",
         "//libs-scala/resources-grpc",
         "@maven//:ch_qos_logback_logback_classic",
-        "@maven//:ch_qos_logback_logback_core",
         "@maven//:com_auth0_java_jwt",
         "@maven//:io_dropwizard_metrics_metrics_core",
         "@maven//:io_dropwizard_metrics_metrics_jmx",
         "@maven//:io_netty_netty_handler",
-        "@maven//:org_slf4j_slf4j_api",
     ],
 )
