+++ ledger/sandbox/src/main/scala/platform/sandboxnext/Main.scala
@@ -4,7 +4,7 @@
 package com.daml.platform.sandboxnext
 
 import com.daml.ledger.resources.ResourceContext
-import com.daml.platform.sandbox.GlobalLogLevel
+import com.daml.cliopts.GlobalLogLevel
 import com.daml.resources.ProgramResource
 
 object Main {
@@ -15,7 +15,7 @@ object Main {
     // [CoordinatedShutdown(akka://sandbox)] Could not addJvmShutdownHook, due to: Shutdown in progress
     System.setProperty("akka.jvm-shutdown-hooks", "off")
     val config = Cli.parse(args).getOrElse(sys.exit(1))
-    config.logLevel.foreach(GlobalLogLevel.set)
+    config.logLevel.foreach(GlobalLogLevel.set("Sandbox"))
     new ProgramResource(new Runner(config)).run(ResourceContext.apply)
   }
